+++ go/falcon-core/physics/device-structures/impedances/impedances_test.go
@@ -39,14 +39,14 @@ func TestImpedances_SizeAndItems(t *testing.T) {
 		if err != nil {
 			t.Fatalf("Size() error: %v", err)
 		}
-		if sz != uint32(len(imps)) {
+		if sz != uint64(len(imps)) {
 			t.Errorf("Size() = %v, want %v", sz, len(imps))
 		}
 		items, err := h.Items()
 		if err != nil {
 			t.Fatalf("Items() error: %v", err)
 		}
-		if size, _ := items.Size(); size != uint32(len(imps)) {
+		if size, _ := items.Size(); size != uint64(len(imps)) {
 			t.Errorf("Items() length = %v, want %v", size, len(imps))
 		}
 	})
@@ -55,7 +55,7 @@ func TestImpedances_SizeAndItems(t *testing.T) {
 func TestImpedances_At(t *testing.T) {
 	withImpedances(t, func(t *testing.T, h *Handle, imps []*impedance.Handle) {
 		for i, want := range imps {
-			got, err := h.At(uint32(i))
+			got, err := h.At(uint64(i))
 			if err != nil {
 				t.Fatalf("At(%d) error: %v", i, err)
 			}
@@ -81,7 +81,7 @@ func TestImpedances_ContainsAndIndex(t *testing.T) {
 			if err != nil {
 				t.Fatalf("Index(%d) error: %v", i, err)
 			}
-			if idx != uint32(i) {
+			if idx != uint64(i) {
 				t.Errorf("Index(%d) = %v, want %v", i, idx, i)
 			}
 		}
@@ -102,7 +102,7 @@ func TestImpedances_PushBack(t *testing.T) {
 			t.Fatalf("PushBack error: %v", err)
 		}
 		sz, _ := h.Size()
-		if sz != uint32(len(imps)+1) {
+		if sz != uint64(len(imps)+1) {
 			t.Errorf("After PushBack, Size() = %v, want %v", sz, len(imps)+1)
 		}
 	})
@@ -119,7 +119,7 @@ func TestImpedances_EraseAtAndClear(t *testing.T) {
 		t.Fatalf("EraseAt error: %v", err)
 	}
 	sz, _ := h.Size()
-	if sz != uint32(len(imps)-1) {
+	if sz != uint64(len(imps)-1) {
 		t.Errorf("After EraseAt, Size() = %v, want %v", sz, len(imps)-1)
 	}
 	if err := h.Clear(); err != nil {
@@ -147,7 +147,7 @@ func TestImpedances_Intersection(t *testing.T) {
 		if err != nil {
 			t.Fatalf("Intersection Items error: %v", err)
 		}
-		if size, _ := items.Size(); size != uint32(len(imps)) {
+		if size, _ := items.Size(); size != uint64(len(imps)) {
 			t.Errorf("Intersection Items = %v, want %v", size, len(imps))
 		}
 	})
